完善了通信；重新调了车轮PID

+++ MasterChip/CODE/mecanum_chassis.h
@@ -40,7 +40,7 @@ typedef struct BaseChassis_t // 底盘抽象结构体
 	PosMode_e pos_mode;	  // 坐标模式：相对/绝对
 	PostureStatus_t PostureStatus;
 	Motor_t motor[4];		  // 驱动电机，编号为左上角开始顺时针0123
-	uint32_t cam_servo_duty;  // 摄像头舵机占空比
+	uint16 cam_servo_duty;  // 摄像头舵机占空比
 	char send_ctrl_msg_flag;  // 控制发送频率
 	char motor_self_check_ok; // 电机自检完成
